ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
transformers==4.35.2
torchvision==0.16.0
pdf2image==1.16.3
pdfminer.six==20221105
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
joblib==1.3.2
numpy==1.24.3